Use Capitano & introduce a preview mode
