+++ theme/package.json
@@ -12,7 +12,7 @@
     "start": "react-static start",
     "stage": "react-static build --staging",
     "build": "react-static build",
-    "serve": "serve dist -p 3000"
+    "serve": "serve dist -p 1337"
   },
   "dependencies": {
     "circular-std": "^1.0.0",
